LayoutTests/media/video-buffered-expected.txt

TEST(video.buffered.length == 0) OK
TEST(video.buffered.start(-1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.end(-1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.start(0)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.end(0)) THROWS(DOMException.INDEX_SIZE_ERR) OK
EVENT(load)
TEST(video.buffered.length == 1) OK
TEST(video.buffered.start(0) == 0) OK
TEST(video.buffered.end(0) > 0) OK
TEST(video.buffered.start(-1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.end(-1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.start(1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
TEST(video.buffered.end(1)) THROWS(DOMException.INDEX_SIZE_ERR) OK
END OF TEST
